% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignature.R
\name{addSignatureDF}
\alias{addSignatureDF}
\title{Add signature providid via data.frame}
\usage{
addSignatureDF(obj, df, ID)
}
\arguments{
\item{obj}{geneSign instance}

\item{df}{data.frame with two columns (first column: Gene,
second columns: Direction of alterations: "up" or "down").
columnnames: Gene, Arbitrary_Name}

\item{ID}{Arbitrary ID of signature}
}
\value{
Updated geneSign object
}
\description{
Adds a gene signature, provided as data.frame
with two columns.
}
