% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fract.R
\name{fract}
\alias{fract}
\title{Calculate fraction of similarly altered genes}
\usage{
fract(obj, fun = median, fractCut = c(0.3, 0.7), CI = 1)
}
\arguments{
\item{obj}{singGene instance}

\item{fun}{Function to determine up/downregulation based
on all values (all samples) for each gene}

\item{fractCut}{Fraction cutoffs to classify sample as
not signature associated or signature associated. If
CI is set to 0, samples with fractions below the first
elements of the fractCut vector are set to 0, above the
second element of fractCut, set to 1. Other samples are
classified as NA.}

\item{CI}{Calculate Wilson confidence interval for fractions
and compare upper 95% limits for the lower cutoff (first
element of fractCut) and lower 95% limits for upper cutoffs
(second element of fractCut)}
}
\value{
Updated singGene object
}
\description{
For each gene, over- or underexpression is
determined by comparing each value to the row-wise values
calculated via fun (default: median; median of all samples).
The number of similiarly altered genes compared to the gene
signature is calculated. Fractions (for each signature)
and binary classifications, based on the fractCut and CI
parameters are returned.
}
