% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signClass.R
\name{initialize,signClass-method}
\alias{initialize,signClass-method}
\title{Constructor incClass}
\usage{
\S4method{initialize}{signClass}(.Object, expr = data.frame, ...)
}
\description{
Constructor for new analysis
}
