% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignature.R
\name{loadSignatureMetadata}
\alias{loadSignatureMetadata}
\title{Load Signature Metadata}
\usage{
loadSignatureMetadata(obj, df = NULL)
}
\description{
Data frame with filenames and group assignments
to use for analysis
}
